New scene structure editor
